backend/requirements.txt
flask
flask-cors
python-dotenv
pandas
xgboost
scikit-learn
